requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
mkl-fft==1.0.12
mkl-random==1.0.2
numpy==1.17.0
Pillow==6.1.0
protobuf==3.9.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.3
Werkzeug==0.15.5
wincertstore==0.2
wrapt==1.11.2
